Add `local-repo-path` for fully local builds
Make it possible to preserve the built wheels for local use, e.g. when
it's needed for subsequent CI steps but publishing the wheels is not
desirable yet. The rsync upload plugin is still internally used to copy
the repository content to a folder in the container, which is then
copied to the path specified in the action input.